+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - Added `Edge.markerStart`
 - Added `data-line-attachement-point` to marker templates to specify where an edge attaches to the marker if the marker is an end marker
+- Added dynamic templates for Nodes, Markers and LinkHandles
+- Added dynamic template registry
 
 ### Deprecated